# trading-chats-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "凌期AI辅助期货挑选品种和持仓止盈止损分析程序API文档",
        "title": "凌期AI辅助期货挑选品种和持仓止盈止损分析程序",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/model-api-configs": {
            "get": {
                "description": "获取所有模型与API配置列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "模型与API配置"
                ],
                "summary": "获取所有模型与API配置",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "创建新的模型与API配置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "模型与API配置"
                ],
                "summary": "创建模型与API配置",
                "parameters": [
                    {
                        "description": "模型与API配置信息",
                        "name": "config",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ModelAPIConfig"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/model-api-configs/provider": {
            "get": {
                "description": "根据提供商获取模型与API配置列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "模型与API配置"
                ],
                "summary": "根据提供商获取模型与API配置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "提供商",
                        "name": "provider",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/model-api-configs/{id}": {
            "get": {
                "description": "根据ID获取模型与API配置详情",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "模型与API配置"
                ],
                "summary": "根据ID获取模型与API配置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "模型与API配置ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            },
            "put": {
                "description": "更新模型与API配置信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "模型与API配置"
                ],
                "summary": "更新模型与API配置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "模型与API配置ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "模型与API配置信息",
                        "name": "config",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ModelAPIConfig"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除模型与API配置",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "模型与API配置"
                ],
                "summary": "删除模型与API配置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "模型与API配置ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/model-api-configs/{id}/test": {
            "post": {
                "description": "测试模型与API的连通性",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "模型与API配置"
                ],
                "summary": "测试模型的连通性",
                "parameters": [
                    {
                        "type": "string",
                        "description": "模型与API配置ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.ModelAPIConfig": {
            "type": "object",
            "properties": {
                "api_key": {
                    "type": "string"
                },
                "api_url": {
                    "type": "string"
                },
                "created_at": {},
                "id": {
                    "type": "string"
                },
                "models": {
                    "description": "一次可添加多个模型名称",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "provider": {
                    "description": "anthropic, openai",
                    "type": "string"
                },
                "updated_at": {}
            }
        },
...
        "models.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
    }
}